feat(trusty-audit): chain install, materialize, collect and package into one command (#5824)
Adds `Command::Audit` / `taudit audit`, which drives the four existing verbs
in order over one working directory. Each phase calls the existing capability
unchanged, so the chain and the separate verbs cannot drift.

The registry (#5822) reaches the sweep for the first time: the materialize
phase clones the registered repository targets, and `clone_all` records the
usable checkouts as the selection `run` reads. A registered board becomes a
stated gap rather than silence.

Splits rendering out of cli.rs into cli/render.rs — the new outcome arm pushed
the file past the 500-SLOC production cap.
